+++ src/main/java/com/idea5/four_cut_photos_map/domain/crawler/service/CrawlService.java
@@ -1,4 +1,4 @@
-package com.idea5.four_cut_photos_map.domain.crawler.service.mainbrand;
+package com.idea5.four_cut_photos_map.domain.crawler.service;
 
 import org.springframework.stereotype.Service;
